feat: Add support for dispatching messages to states
Initially I implemented passing messages as a reference:

  fn initial(&mut self, msg: &Messages) -> StateResult

But then I thought it would be nice to allow other types such as:

  fn initial(&mut self, msg: Option<&Messages>) -> StateResult

When trying to add support for Option<xx> I had to implement the "msg_type"
code differently. I then realized I might be able to handle both
techniques so that's what has been implemented using:

 enum MsgType {
     MtTypePath { tp: syn::TypePath },
     MtTypeReference { tr: syn::TypeReference },
 }

Currently these seem to be working, no guarantees :)

 * TypePath: MsgTy, Option<MsgTy>, Option<&MsgTy>
 * TypeReference: &MsgTy and &mut MsgTy

I'll be adding tests in the near future as to provide some level of
guarantees to what is supported.