feat: Add hsm1_initial_state
This is a proc_macro_attribute which is used to identify the initial
Previously the initial state had to have "initial" as its name. Forcing
the name to be initial wasn't descriptive enough, especially when the
initial state is used as part of the typical state sequence.

Another alternative was to require the "initial" state to have two
proc_maro_attributes, but I like this technique although I might
change my mind in the furture.