refactor: Change StateResult to a tuple istead of an enum
This allows less code duplication in Executor and makes it more explicit,
for good or evil, that the user is returning two items, Handled and
Transition.

This does mean there is more typing in the simple no-transition cases
so we'll see if I stick with this in the long term.