feat: Add dispatcher
Also require Debug for SM and P to make debugging easier. This will be
a "feature" in the future.

Use dispatcher instead of dispatch, although calling dispatch would be
an optimization if it's known that the SM does not use defer_send.

Changed test_sm_1s_no_enter_no_exit to use not have sme as member of
StateMacine since we now pass sme to ProcessFn's.