bug: Change dispatcher to process all deferred messages
This needs to be done so that we can guarantee after a transition we
process "older" messages before "newer" messages. This can be dangerous
so care needs to be taken maybe some other technique found to handle
situations where this is uesful.

Also, Fix bug in next_defer and changed to use log::trace.