feat: Executor.sm does not need to be Rc<RefCell<SM>>
I only needs to be a RefCell<SM>. This is required because we
pass &mut sm and &Self (i.e. Executor) and otherwise we get
the following error at the call site in dispatch_idx:

  (self.states[idx].process)(&mut self.sm, self, msg);
  -------------------------- ------------  ^^^^ immutable borrow occurs here
  |                          |
  |                          mutable borrow occurs here
  mutable borrow later used by call

This also means the StateMachine instance is in Executor and on the stack
rather than in the heap when it was an Rc. And thus the size change of
get_sm() from 8, when it was an Rc. To 16 now that it's a RefCell.

 $ cargo run
    Compiling hsm0-executor v0.6.0 (/home/wink/prgs/rust/myrepos/proc-macro-hsm1/hsm0-executor)
     Finished dev [unoptimized + debuginfo] target(s) in 0.48s
      Running `/home/wink/prgs/rust/myrepos/proc-macro-hsm1/target/debug/hsm0-executor`
 main:+
 main:  &sme=0x7ffdd51d7e68
 state1:+ &self=0x7ffdd51d7e70
 state1:-
 state1:+ &self=0x7ffdd51d7e70
 state2:-
 state1:+ &self=0x7ffdd51d7e70
 state1:-
 state1:+ &self=0x7ffdd51d7e70
 state2:-
 main:-